fix type of foreign keys in migrations

+++ backend/database/migrations/2022_12_11_231157_create_sale_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('sale', function (Blueprint $table) {
             $table->id();
             $table->float("total_value");
-            $table->integer("id_client");
+            $table->unsignedBigInteger("id_client");
             $table->timestamps();
 
             $table->foreign('id_client')->references('id')->on('client');
